refactor(domain-pack): draft component 저장 책임을 분리
