+++ backend/src/main/java/com/init/domainpack/application/DraftComponentsInput.java
@@ -1,0 +1,10 @@
+package com.init.domainpack.application;
+
+import java.util.List;
+
+record DraftComponentsInput(
+    List<SlotInput> slots,
+    List<PolicyInput> policies,
+    List<RiskInput> risks,
+    List<WorkflowInput> workflows,
+    List<IntentSlotBindingInput> intentSlotBindings) {}
