+++ backend/src/main/java/com/init/domainpack/application/DraftPayload.java
@@ -1,0 +1,11 @@
+package com.init.domainpack.application;
+
+import java.util.List;
+
+record DraftPayload(
+    List<IntentDraft> intents,
+    List<SlotInput> slots,
+    List<IntentSlotBindingInput> intentSlotBindings,
+    List<PolicyInput> policies,
+    List<RiskInput> risks,
+    List<WorkflowInput> workflows) {}
