+++ backend/src/main/java/com/init/domainpack/application/IntentSlotBindingInput.java
@@ -1,0 +1,9 @@
+package com.init.domainpack.application;
+
+record IntentSlotBindingInput(
+    String intentCode,
+    String slotCode,
+    Boolean isRequired,
+    Integer collectionOrder,
+    String promptHint,
+    String conditionJson) {}
