+++ backend/src/main/java/com/init/domainpack/application/PersistedDraftIntents.java
@@ -1,0 +1,7 @@
+package com.init.domainpack.application;
+
+import com.init.domainpack.domain.model.IntentDefinition;
+import java.util.List;
+
+record PersistedDraftIntents(
+    List<IntentDefinition> savedIntents, int skippedIntentCount, int totalIntentCount) {}
