+++ backend/src/main/java/com/init/domainpack/application/PolicyInput.java
@@ -1,0 +1,11 @@
+package com.init.domainpack.application;
+
+record PolicyInput(
+    String policyCode,
+    String name,
+    String description,
+    String severity,
+    String conditionJson,
+    String actionJson,
+    String evidenceJson,
+    String metaJson) {}
