+++ backend/src/main/java/com/init/domainpack/application/RiskInput.java
@@ -1,0 +1,11 @@
+package com.init.domainpack.application;
+
+record RiskInput(
+    String riskCode,
+    String name,
+    String description,
+    String riskLevel,
+    String triggerConditionJson,
+    String handlingActionJson,
+    String evidenceJson,
+    String metaJson) {}
