+++ backend/src/main/java/com/init/domainpack/application/SavedDraftComponents.java
@@ -1,0 +1,8 @@
+package com.init.domainpack.application;
+
+record SavedDraftComponents(
+    int addedSlotCount,
+    int addedPolicyCount,
+    int addedRiskCount,
+    int addedWorkflowCount,
+    int addedIntentSlotBindingCount) {}
