+++ backend/src/main/java/com/init/domainpack/application/SlotInput.java
@@ -1,0 +1,11 @@
+package com.init.domainpack.application;
+
+record SlotInput(
+    String slotCode,
+    String name,
+    String description,
+    String dataType,
+    Boolean isSensitive,
+    String validationRuleJson,
+    String defaultValueJson,
+    String metaJson) {}
